+++ app/messages/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import PlaybookInbox from "@/components/messages/PlaybookInbox";
+
+export default function MessagesPage() {
+  return <PlaybookInbox />;
+}